LayoutTests/imported/w3c/web-platform-tests/websockets/Send-data.any-expected.txt


FAIL Send data on a WebSocket - Connection should be opened assert_equals: expected 0 but got 15
PASS Send data on a WebSocket - Message should be received
FAIL Send data on a WebSocket - Connection should be closed assert_true: WebSocket connection should be open expected true got false
